refactor(nexus-web): migrate Files to Account-style UI module
Colocate browse route, mobile components, route parser, and semantic CSS under files/ per AGENTS.md Phase 1.
