+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/app/workspace/[workspaceId]/files/browse/browse.css
@@ -1,0 +1,80 @@
+/* files/browse — route layout and responsive visibility */
+
+.files-browse-root {
+  display: flex;
+  height: 100%;
+  flex-direction: column;
+}
+
+@media (max-width: 767px) {
+  .files-browse-root .files-toolbar-desktop {
+    display: none;
+  }
+
+  .files-browse-root .files-table-view {
+    display: none;
+  }
+
+  .files-browse-root .files-mobile-list {
+    display: flex;
+  }
+
+  .files-browse-root .files-pagination-compact .files-pagination-page-size {
+    display: none;
+  }
+
+  .files-browse-root .files-pagination-compact {
+    flex-direction: column;
+    align-items: stretch;
+    gap: var(--space-2);
+    padding-left: var(--space-4);
+    padding-right: var(--space-4);
+  }
+
+  .files-browse-root .files-pagination-compact .files-pagination-controls {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    width: 100%;
+    gap: var(--space-2);
+  }
+
+  .files-browse-root .files-content-area {
+    padding: 0;
+  }
+
+  .files-browse-root .files-breadcrumb {
+    padding-left: var(--space-4);
+    padding-right: var(--space-4);
+    overflow-x: auto;
+    white-space: nowrap;
+    -webkit-overflow-scrolling: touch;
+  }
+}
+
+@media (min-width: 768px) {
+  .files-browse-root .files-mobile-toolbar {
+    display: none;
+  }
+
+  .files-browse-root .files-mobile-list {
+    display: none;
+  }
+}
+
+.files-browse-root .files-breadcrumb {
+  display: flex;
+  align-items: center;
+  gap: var(--space-1);
+  border-bottom: 1px solid var(--border-hex);
+  padding: var(--space-2) var(--space-4);
+  font-size: var(--font-size-sm, 14px);
+}
+
+.files-browse-root .files-content-area.is-dragging {
+  background: var(--color-primary-subtle);
+}
+
+.files-browse-root .files-mobile-list {
+  display: none;
+}
