+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/app/workspace/[workspaceId]/files/browse/page.tsx
@@ -1,1 +1,1 @@
-export { default } from '../page';
+export { default } from './browse';
