+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/app/workspace/[workspaceId]/files/components/files-components.css
@@ -1,0 +1,283 @@
+/* files/components — shared semantic styles for mobile chrome */
+
+.files-mobile-toolbar {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-2);
+  padding: var(--space-2) var(--space-4);
+}
+
+.files-mobile-toolbar-row {
+  display: flex;
+  align-items: center;
+  gap: var(--space-2);
+  width: 100%;
+}
+
+.files-mobile-toolbar-add {
+  display: inline-flex;
+  flex-shrink: 0;
+  align-items: center;
+  gap: 6px;
+  height: 36px;
+  padding: 0 var(--space-3);
+  border: none;
+  border-radius: var(--org-border-radius, var(--radius-md));
+  font-size: var(--font-size-sm, 14px);
+  font-weight: var(--font-weight-medium, 500);
+  background: var(--color-primary);
+  color: var(--color-primary-foreground);
+  cursor: pointer;
+}
+
+.files-mobile-toolbar-icon-btn {
+  display: inline-flex;
+  flex-shrink: 0;
+  align-items: center;
+  justify-content: center;
+  height: 36px;
+  width: 36px;
+  border-radius: var(--org-border-radius, var(--radius-md));
+  border: 1px solid var(--border-hex);
+  background: transparent;
+  cursor: pointer;
+}
+
+.files-mobile-toolbar-icon-btn.is-disabled {
+  opacity: 0.5;
+  pointer-events: none;
+}
+
+.files-mobile-toolbar-spacer {
+  flex: 1;
+  min-width: 0;
+}
+
+.files-mobile-toolbar-search {
+  position: relative;
+  width: 100%;
+}
+
+.files-mobile-toolbar-search input {
+  width: 100%;
+  height: 36px;
+  padding-left: 36px;
+  padding-right: var(--space-3);
+  border-radius: var(--org-border-radius, var(--radius-md));
+  border: 1px solid var(--border-hex);
+  background: transparent;
+  font-size: var(--font-size-sm, 14px);
+  outline: none;
+}
+
+.files-mobile-toolbar-search input:focus {
+  box-shadow: var(--shadow-focus);
+}
+
+.files-mobile-toolbar-search-icon {
+  position: absolute;
+  left: 10px;
+  top: 50%;
+  transform: translateY(-50%);
+  color: var(--muted-foreground-hex);
+  pointer-events: none;
+}
+
+.files-mobile-toolbar-menu {
+  position: absolute;
+  right: 0;
+  top: 100%;
+  z-index: 50;
+  margin-top: var(--space-1);
+  min-width: 160px;
+  border-radius: var(--org-border-radius, var(--radius-md));
+  border: 1px solid var(--border-hex);
+  background: var(--card-hex);
+  padding: var(--space-1) 0;
+  box-shadow: var(--glass-shadow);
+}
+
+.files-mobile-toolbar-menu-item {
+  display: flex;
+  width: 100%;
+  align-items: center;
+  gap: var(--space-2);
+  padding: var(--space-2) var(--space-3);
+  border: none;
+  background: transparent;
+  font-size: var(--font-size-sm, 14px);
+  text-align: left;
+  cursor: pointer;
+}
+
+.files-mobile-toolbar-menu-item:hover {
+  background: var(--muted-bg-hex);
+}
+
+.files-mobile-toolbar-menu-item.is-active {
+  background: var(--surface-subtle-hex);
+}
+
+/* Mobile file rows */
+.files-mobile-list {
+  display: flex;
+  flex-direction: column;
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
+
+.files-mobile-row {
+  display: flex;
+  align-items: center;
+  gap: var(--space-1);
+  min-height: 48px;
+  padding-inline: var(--space-4);
+  border-bottom: 1px solid var(--border-hex);
+}
+
+.files-mobile-row.is-drop-target {
+  background: var(--color-primary-subtle);
+  box-shadow: inset 0 0 0 1px var(--color-primary);
+}
+
+.files-mobile-row-main {
+  display: flex;
+  flex: 1;
+  min-width: 0;
+  align-items: center;
+  gap: var(--space-3);
+  padding: var(--space-2) 0;
+  border: none;
+  background: transparent;
+  text-align: left;
+  font-size: var(--font-size-sm, 14px);
+  color: var(--foreground-hex);
+  cursor: pointer;
+}
+
+.files-mobile-row-main:hover {
+  color: var(--color-primary);
+}
+
+.files-mobile-row-icon {
+  flex-shrink: 0;
+}
+
+.files-mobile-row-text {
+  flex: 1;
+  min-width: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+}
+
+.files-mobile-row-name {
+  line-height: 20px;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.files-mobile-row-meta {
+  font-size: 12px;
+  line-height: 16px;
+  color: var(--muted-foreground-hex);
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.files-mobile-row-actions {
+  flex-shrink: 0;
+}
+
+.files-mobile-row-actions button[title='More options'],
+.files-mobile-row-actions > div > button:last-child {
+  opacity: 1;
+}
+
+/* Add bottom sheet (mobile) */
+.files-add-sheet-backdrop {
+  position: fixed;
+  inset: 0;
+  z-index: 250;
+  border: none;
+  background: rgb(0 0 0 / 0.4);
+  cursor: pointer;
+}
+
+.files-add-sheet-panel {
+  position: fixed;
+  inset-inline: 0;
+  bottom: 0;
+  z-index: 251;
+  border-top: 1px solid var(--border-hex);
+  background: var(--background-hex);
+  box-shadow: 0 -4px 24px rgb(0 0 0 / 0.12);
+  padding-bottom: env(safe-area-inset-bottom, 0px);
+  animation: files-add-sheet-in 0.2s ease-out;
+}
+
+@keyframes files-add-sheet-in {
+  from {
+    transform: translateY(100%);
+  }
+  to {
+    transform: translateY(0);
+  }
+}
+
+.files-add-sheet-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: var(--space-3) var(--space-4) var(--space-2);
+}
+
+.files-add-sheet-header h2 {
+  margin: 0;
+  font-size: var(--font-size-sm, 14px);
+  font-weight: var(--font-weight-semibold, 600);
+}
+
+.files-add-sheet-close {
+  display: flex;
+  height: 32px;
+  width: 32px;
+  align-items: center;
+  justify-content: center;
+  border: none;
+  border-radius: var(--org-border-radius, var(--radius-md));
+  background: transparent;
+  color: var(--muted-foreground-hex);
+  cursor: pointer;
+}
+
+.files-add-sheet-close:hover {
+  background: var(--muted-bg-hex);
+}
+
+.files-add-sheet-actions {
+  display: flex;
+  flex-direction: column;
+  padding: 0 var(--space-2) var(--space-3);
+}
+
+.files-add-sheet-actions button {
+  display: flex;
+  align-items: center;
+  gap: var(--space-3);
+  width: 100%;
+  padding: 14px var(--space-3);
+  border: none;
+  border-radius: var(--org-border-radius, var(--radius-md));
+  background: transparent;
+  font-size: 15px;
+  text-align: left;
+  cursor: pointer;
+}
+
+.files-add-sheet-actions button:hover {
+  background: var(--surface-subtle-hex);
+}
