--- libs/naas-abi/naas_abi/apps/nexus/apps/web/src/app/workspace/[workspaceId]/files/files-page.css
@@ -1,260 +1,0 @@
-/* Mobile files browse — OneDrive-style list rows and compact chrome */
-
-@media (max-width: 767px) {
-  .files-browse-root .files-toolbar-desktop {
-    display: none;
-  }
-
-  .files-browse-root .files-table-view {
-    display: none;
-  }
-
-  .files-browse-root .files-mobile-list {
-    display: flex;
-    flex-direction: column;
-    list-style: none;
-    margin: 0;
-    padding: 0;
-  }
-
-  .files-browse-root .files-mobile-row {
-    display: flex;
-    align-items: center;
-    gap: 0.25rem;
-    min-height: 3rem;
-    padding-inline: var(--space-4);
-    border-bottom: 1px solid hsl(var(--border) / 0.5);
-  }
-
-  .files-browse-root .files-mobile-row.is-drop-target {
-    background: hsl(var(--primary) / 0.1);
-    box-shadow: inset 0 0 0 1px hsl(var(--primary));
-  }
-
-  .files-browse-root .files-mobile-row-main {
-    display: flex;
-    flex: 1;
-    min-width: 0;
-    align-items: center;
-    gap: 0.75rem;
-    padding: 0.5rem 0;
-    text-align: left;
-    font-size: 0.875rem;
-  }
-
-  .files-browse-root .files-mobile-row-main:hover {
-    color: hsl(var(--primary));
-  }
-
-  .files-browse-root .files-mobile-row-icon {
-    flex-shrink: 0;
-  }
-
-  .files-browse-root .files-mobile-row-text {
-    flex: 1;
-    min-width: 0;
-    display: flex;
-    flex-direction: column;
-    gap: 0.125rem;
-  }
-
-  .files-browse-root .files-mobile-row-name {
-    line-height: 1.25rem;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-
-  .files-browse-root .files-mobile-row-meta {
-    font-size: 0.75rem;
-    line-height: 1rem;
-    color: hsl(var(--muted-foreground));
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-
-  .files-browse-root .files-mobile-row-actions {
-    flex-shrink: 0;
-  }
-
-  .files-browse-root .files-mobile-row-actions button[title='More options'],
-  .files-browse-root .files-mobile-row-actions > div > button:last-child {
-    opacity: 1;
-  }
-
-  .files-browse-root .files-toolbar-mobile {
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-    padding: 0.5rem var(--space-4);
-  }
-
-  .files-browse-root .files-toolbar-mobile-row {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    width: 100%;
-  }
-
-  .files-browse-root .files-toolbar-mobile-add {
-    display: inline-flex;
-    flex-shrink: 0;
-    align-items: center;
-    gap: 0.375rem;
-    height: 2.25rem;
-    padding: 0 0.75rem;
-    border-radius: 0.375rem;
-    font-size: 0.875rem;
-    font-weight: 500;
-    background: hsl(var(--primary));
-    color: hsl(var(--primary-foreground));
-  }
-
-  .files-browse-root .files-toolbar-mobile-icon-btn {
-    display: inline-flex;
-    flex-shrink: 0;
-    align-items: center;
-    justify-content: center;
-    height: 2.25rem;
-    width: 2.25rem;
-    border-radius: 0.375rem;
-    border: 1px solid hsl(var(--border));
-  }
-
-  .files-browse-root .files-toolbar-mobile-spacer {
-    flex: 1;
-    min-width: 0;
-  }
-
-  .files-browse-root .files-toolbar-mobile-search {
-    position: relative;
-    width: 100%;
-  }
-
-  .files-browse-root .files-toolbar-mobile-search input {
-    width: 100%;
-    height: 2.25rem;
-    padding-left: 2.25rem;
-    padding-right: 0.75rem;
-    border-radius: 0.375rem;
-    border: 1px solid hsl(var(--border));
-    background: transparent;
-    font-size: 0.875rem;
-    outline: none;
-  }
-
-  .files-browse-root .files-toolbar-mobile-search svg {
-    position: absolute;
-    left: 0.625rem;
-    top: 50%;
-    transform: translateY(-50%);
-    color: hsl(var(--muted-foreground));
-    pointer-events: none;
-  }
-
-  .files-browse-root .files-pagination-compact .files-pagination-page-size {
-    display: none;
-  }
-
-  .files-browse-root .files-pagination-compact {
-    flex-direction: column;
-    align-items: stretch;
-    gap: 0.5rem;
-    padding-left: var(--space-4);
-    padding-right: var(--space-4);
-  }
-
-  .files-browse-root .files-pagination-compact .files-pagination-controls {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-    gap: 0.5rem;
-  }
-
-  .files-browse-root .files-content-area {
-    padding: 0;
-  }
-
-  .files-browse-root .files-breadcrumb {
-    padding-left: var(--space-4);
-    padding-right: var(--space-4);
-    overflow-x: auto;
-    white-space: nowrap;
-    -webkit-overflow-scrolling: touch;
-  }
-}
-
-@media (min-width: 768px) {
-  .files-browse-root .files-toolbar-mobile {
-    display: none;
-  }
-
-  .files-browse-root .files-mobile-list {
-    display: none;
-  }
-}
-
-/* Add bottom sheet (mobile) */
-.files-add-sheet-backdrop {
-  position: fixed;
-  inset: 0;
-  z-index: 250;
-  background: rgb(0 0 0 / 0.4);
-}
-
-.files-add-sheet-panel {
-  position: fixed;
-  inset-inline: 0;
-  bottom: 0;
-  z-index: 251;
-  border-top: 1px solid hsl(var(--border) / 0.6);
-  background: hsl(var(--background));
-  box-shadow: 0 -4px 24px rgb(0 0 0 / 0.12);
-  padding-bottom: env(safe-area-inset-bottom, 0px);
-  animation: files-add-sheet-in 0.2s ease-out;
-}
-
-@keyframes files-add-sheet-in {
-  from {
-    transform: translateY(100%);
-  }
-  to {
-    transform: translateY(0);
-  }
-}
-
-.files-add-sheet-header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0.75rem 1rem 0.5rem;
-}
-
-.files-add-sheet-header h2 {
-  margin: 0;
-  font-size: 0.875rem;
-  font-weight: 600;
-}
-
-.files-add-sheet-actions {
-  display: flex;
-  flex-direction: column;
-  padding: 0 0.5rem 0.75rem;
-}
-
-.files-add-sheet-actions button {
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-  width: 100%;
-  padding: 0.875rem 0.75rem;
-  border-radius: 0.375rem;
-  font-size: 0.9375rem;
-  text-align: left;
-}
-
-.files-add-sheet-actions button:hover {
-  background: hsl(var(--muted) / 0.5);
-}
